@@ -15,10 +15,10 @@ public class ErrorViewModel : DotvvmViewModelBase
     public string? RequestId { get; set; }
 
     [Bind(Direction.None)]
-    public string ExceptionType { get; set; }
+    public string? ExceptionType { get; set; }
 
     [Bind(Direction.None)]
-    public string RequestPath { get; set; }
+    public string? RequestPath { get; set; }
 
 
     public ErrorViewModel()
@@ -29,6 +29,11 @@ public class ErrorViewModel : DotvvmViewModelBase
     {
         var aspcontext = Context.GetAspNetCoreContext();
         var exceptionInfo = aspcontext.Features.Get<IExceptionHandlerFeature>();
+        if (exceptionInfo is null)
+        {
+            ExceptionType = "View called without IExceptionHandlerFeature";
+            return base.Init();
+        }
         ExceptionType = exceptionInfo.Error.GetType().Name;
         RequestId = aspcontext.TraceIdentifier;
         RequestPath = exceptionInfo.Path;
